Readjust screen size to a more portrait ratio
Still having to fix the screen size which is not great, and it doesn't
seem to be scaling well. Will need to keep tweaking how we can calculate
positioning to allow better scaling but this will do for now. Just make
sure to refactor everything really well as we anticipate more changes in
the future.

Also change the data flow for battle control slightly as we are now only
passing player's party in since we don't need controls for the opponents
anyway.

Also refactored the menu drawing a bit.